refactor: 패키지 분리, 이름 변경
