+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.domain.CarRacingManager;
+import racingcar.controller.RacingController;
 
 import java.util.Scanner;
 
@@ -8,7 +8,7 @@ public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
 
-        CarRacingManager carRacingManager = new CarRacingManager(scanner);
+        RacingController carRacingManager = new RacingController(scanner);
         carRacingManager.start();
     }
 }
